# ai-specs/specs/api-spec.yml
openapi: 3.0.0
info:
  title: DocFlow API
  description: |
    REST API for DocFlow — Intelligent Document Management System (DMS).
    
    DocFlow is a modular, API-First document management infrastructure designed to serve as a central repository
    for advanced document lifecycle management, versioning, and semantic search capabilities.
    
    The platform implements a **microservices architecture** with the following core domains:
    - **Identity & Authentication**: Multi-organization support with JWT-based stateless authentication (Identity Service)
    - **Document Management**: Core document lifecycle management, versioning, and storage (Document-Core Service)
    - **API Gateway**: Single entry point routing to backend microservices (Gateway Service)
    - **Audit & Compliance**: Activity tracking and audit logging (AuditLog Service)
    
    ## Architecture & Design Principles
    - **Hexagonal Architecture** (Ports & Adapters): Clear separation between domain logic, application services, and infrastructure
    - **Domain-Driven Design (DDD)**: Business-centric modeling with ubiquitous language
    - **RBAC (Role-Based Access Control)**: Fine-grained permission management across resources
    - **Stateless Authentication**: JWT tokens with organization context
    
    ## Tech Stack
    - **Backend**: Java 21, Spring Boot 3.4.2+, Spring Cloud Gateway, Spring WebFlux
    - **Frontend**: React 19+, Vite, TypeScript, Tailwind CSS, Zustand (state management)
...
    
...
servers:
...
paths:
...
  /api/acl/niveles:
    get:
      summary: List active access levels
      description: Returns all active access levels (LECTURA, ESCRITURA, ADMINISTRACION) ordered by 'orden' field for use in permission management interfaces
      tags:
        - ACL - Access Levels
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of active access levels retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NivelAccesoDTO'
        '401':
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/acl/niveles/codigo/{codigo}:
    get:
      summary: Get access level by codigo
      description: Retrieve a specific access level by its invariable code (LECTURA, ESCRITURA, or ADMINISTRACION)
      tags:
        - ACL - Access Levels
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: codigo
          required: true
          schema:
            type: string
            enum: [LECTURA, ESCRITURA, ADMINISTRACION]
          description: Access level code
      responses:
        '200':
          description: Access level retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NivelAccesoDTO'
        '400':
          description: Bad request - invalid or unknown codigo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/carpetas/{id}:
    delete:
      summary: Delete empty folder (soft delete)
      description: |
        Performs a logical deletion of an empty folder. The folder is not physically removed.
        Requires ADMINISTRACION permission and the folder must have no active subfolders or documents.
        Root folders cannot be deleted.
      tags:
        - Folders
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Folder ID to delete
          example: 120
        - name: X-User-Id
          in: header
          required: true
          schema:
            type: integer
            format: int64
          description: Authenticated user ID (injected by gateway)
          example: 1
        - name: X-Organization-Id
          in: header
          required: true
          schema:
            type: integer
            format: int64
          description: User's organization ID (injected by gateway)
          example: 1
      responses:
        '204':
          description: Folder deleted successfully
        '400':
          description: Root folder cannot be deleted
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Folder not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Folder is not empty
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/carpetas/{carpetaId}/permisos:
    post:
      summary: Grant folder permission to user
      description: Creates an explicit ACL for a user over a folder. Requires ADMINISTRACION.
      tags:
        - ACL - Folder Permissions
      security:
        - bearerAuth: []
      parameters:
        - name: carpetaId
          in: path
          required: true
          schema:
            type: integer
          description: Folder ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePermisoCarpetaUsuarioRequest'
      responses:
        '201':
          description: Permission created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermisoCarpetaUsuarioResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Folder or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate permission
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List folder permissions
      description: Lists all explicit user permissions for a folder.
      tags:
        - ACL - Folder Permissions
      security:
        - bearerAuth: []
      parameters:
        - name: carpetaId
          in: path
          required: true
          schema:
            type: integer
          description: Folder ID
      responses:
        '200':
          description: List of permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermisoCarpetaUsuarioResponse'
        '404':
          description: Folder not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/carpetas/{carpetaId}/permisos/{usuarioId}:
    patch:
      summary: Update folder permission
      description: Updates the access level or recursion for a user on a folder.
      tags:
        - ACL - Folder Permissions
      security:
        - bearerAuth: []
      parameters:
        - name: carpetaId
          in: path
          required: true
          schema:
            type: integer
          description: Folder ID
        - name: usuarioId
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePermisoCarpetaUsuarioRequest'
      responses:
        '200':
          description: Permission updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermisoCarpetaUsuarioResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Permission not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Revoke folder permission
      description: Revokes a user's explicit ACL over a folder (hard delete).
      tags:
        - ACL - Folder Permissions
      security:
        - bearerAuth: []
      parameters:
        - name: carpetaId
          in: path
          required: true
          schema:
            type: integer
          description: Folder ID
        - name: usuarioId
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '204':
          description: Permission revoked successfully
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Permission, folder, or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Document Permission Management (ACL)
...
  /api/documentos/{documentoId}/permisos/{usuarioId}:
    delete:
      summary: Revoke document permission
      description: Revokes a user's explicit ACL over a document (hard delete)
      tags:
        - ACL - Document Permissions
      security:
        - bearerAuth: []
      parameters:
        - name: documentoId
          in: path
          required: true
          schema:
            type: integer
          description: Document ID
        - name: usuarioId
          in: path
          required: true
          schema:
            type: integer
          description: User ID
      responses:
        '204':
          description: Permission revoked successfully
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Permission, document, or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    NivelAccesoDTO:
      type: object
      description: Access Level Data Transfer Object representing permission levels in RBAC
      properties:
        id:
          type: string
          format: Long
          description: Unique access level identifier
        codigo:
          type: string
          enum: [LECTURA, ESCRITURA, ADMINISTRACION]
          description: Invariable access level code
        nombre:
          type: string
          maxLength: 100
          description: Display name of the access level
        descripcion:
          type: string
          nullable: true
          description: Detailed description of permissions granted by this level
        acciones_permitidas:
          type: array
          items:
            type: string
          description: List of allowed actions (e.g., ["ver", "listar", "descargar"])
          example: ["ver", "listar", "descargar"]
        orden:
          type: integer
          description: Display order for UI listings
        activo:
          type: boolean
          description: Whether this access level is currently active
        fecha_creacion:
          type: string
          format: date-time
          description: When this access level was created
        fecha_actualizacion:
          type: string
          format: date-time
          description: Last time this access level was updated
      required:
        - id
        - codigo
        - nombre
        - acciones_permitidas
        - activo
    CreatePermisoCarpetaUsuarioRequest:
      type: object
      properties:
        usuario_id:
          type: integer
          description: User ID to grant access
        nivel_acceso_codigo:
          type: string
          enum: [LECTURA, ESCRITURA, ADMINISTRACION]
          description: Access level code
        recursivo:
          type: boolean
          description: Applies permission to subfolders
        comentario_opcional:
          type: string
          nullable: true
          description: Optional audit comment
      required:
        - usuario_id
        - nivel_acceso_codigo
        - recursivo
    UpdatePermisoCarpetaUsuarioRequest:
      type: object
      properties:
        nivel_acceso_codigo:
          type: string
          enum: [LECTURA, ESCRITURA, ADMINISTRACION]
          description: New access level code
        recursivo:
          type: boolean
          nullable: true
          description: Update recursive flag
      required:
        - nivel_acceso_codigo
    UsuarioResumenDTO:
      type: object
      properties:
        id:
          type: integer
          description: User ID
        email:
          type: string
          format: email
          nullable: true
          description: User email
        nombre:
          type: string
          nullable: true
          description: User full name
      required:
        - id
    PermisoCarpetaUsuarioResponse:
      type: object
      properties:
        id:
          type: integer
          description: Permission ID
        carpeta_id:
          type: integer
          description: Folder ID
        usuario_id:
          type: integer
          description: User ID
        usuario:
          $ref: '#/components/schemas/UsuarioResumenDTO'
        nivel_acceso:
          $ref: '#/components/schemas/NivelAccesoDTO'
        recursivo:
          type: boolean
          description: Recursive permission flag
        fecha_creacion:
          type: string
          format: date-time
          description: Creation timestamp
        fecha_actualizacion:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - id
        - carpeta_id
        - usuario_id

    # Document Operations Schemas
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code/identifier
        message:
          type: string
          description: Human-readable error message
        timestamp:
          type: string
          format: date-time
          description: When the error occurred
        path:
          type: string
          description: Request path that caused the error
        status:
          type: integer
          description: HTTP status code
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: Additional error details (validation errors, etc.)
      required:
        - error
        - message
        - status
...
security:
  - bearerAuth: []
...